Update plugin to work with Godot 4 Beta 13+
Godot 4 Beta 13 changed the required order of @icon in relation to @class.  Annotations must appear before the thing they annotate.

This corrects the order resolving the issue for GD4 Beta 13, as this is the reverse of the required ordering for GD4 Beta 12 applying this means that the plugin will no longer work for installs of <= GD4 Beta 12,

See the following issues for more details:
- https://github.com/godotengine/godot/issues/71592
- https://github.com/godotengine/godot-docs/issues/6629

The PR that introduces the change:
- https://github.com/godotengine/godot/pull/67774